redirect to sign in if not logged in

+++ src/api/projects/deleteProject.ts
@@ -1,10 +1,15 @@
-import {supabase} from "../../../supabaseClient.ts";
+import { supabaseClient } from "../../../supabaseClient.ts";
 
-export async function deleteProject(projectId:string) {
+export async function deleteProject(id, userId, token) {
+    const supabase = await supabaseClient(token);
+
     const { data, error } = await supabase
-        .from('Projects')
+        .from('projects')
         .delete()
-        .match({ id: projectId });
+        .eq('id', id)
+        .eq('user_id', userId);
+
     if (error) throw error;
+
     return data;
 }
